Payment page added

+++ api/index.tsx
@@ -10,25 +10,25 @@ import {
 // auth
 
 export const onLogin = function (body: SiginFields) {
-  return Api.post("/auth/login", body);
+  return Api.post("/admin/auth/login", body);
 };
 
 export const logout = function (body: {
   refreshToken: string;
   fcmToken?: string;
 }) {
-  return Api.post("/auth/logout", body);
+  return Api.post("/admin/auth/logout", body);
 };
 
 export const forgetPassword = function (body: ForgetFields) {
-  return Api.post("/auth/forget-password", body);
+  return Api.post("/admin/auth/forget-password", body);
 };
 
 export const resetPassword = function (
   body: ResetPasswordType,
   token: string | null
 ) {
-  return axios.post(BASE_URL + "/auth/reset-password", body, {
+  return axios.post(BASE_URL + "/admin/auth/reset-password", body, {
     headers: {
       Authorization: `Bearer ${token}`,
     },
@@ -36,11 +36,11 @@ export const resetPassword = function (
 };
 
 export const changePassword = function (body: ResetPasswordType) {
-  return Api.post("/auth/reset-password", body);
+  return Api.post("/admin/auth/reset-password", body);
 };
 
 export const getStats = function () {
-  return Api.get("/stats");
+  return Api.get("/admin/stats");
 };
 
 export const getUsers = function (
@@ -48,7 +48,7 @@ export const getUsers = function (
   limit: number = 10,
   email?: string
 ) {
-  return Api.get(`/user?page=${page}&limit=${limit}&email=${email || ""}`);
+  return Api.get(`/admin/user?page=${page}&limit=${limit}&email=${email || ""}`);
 };
 
 export const getSubadmin = function (
@@ -56,37 +56,37 @@ export const getSubadmin = function (
   limit: number = 10,
   email?: string
 ) {
-  return Api.get(`/sub-admin?page=${page}&limit=${limit}&email=${email}`);
+  return Api.get(`/admin/sub-admin?page=${page}&limit=${limit}&email=${email}`);
 };
 
 export const onAddSubAdmin = function (body: any) {
-  return Api.post("/sub-admin/create", body);
+  return Api.post("/admin/sub-admin/create", body);
 };
 
 export const getGuidline = function (type: number) {
-  return Api.get(`/guideline?type=${type}`);
+  return Api.get(`/admin/guideline?type=${type}`);
 };
 
 export const addGuidline = function (body: any) {
-  return Api.post("/guideline/create", body);
+  return Api.post("/admin/guideline/create", body);
 };
 
 export const updateGuidline = function (id: string, body: any) {
-  return Api.patch(`/guideline/update/${id}`, body);
+  return Api.patch(`/admin/guideline/update/${id}`, body);
 };
 
 export const getNotification = function (page: number = 1, limit: number = 10) {
-  return Api.get(`/notification?page=${page}&limit=${limit}`);
+  return Api.get(`/admin/notification?page=${page}&limit=${limit}`);
 };
 
 export const sendNotification = function (body: any) {
-  return Api.post("/notification/send", body);
+  return Api.post("/admin/notification/send", body);
 };
 
 export const sendMedia = async (media: any) => {
   let formData = new FormData();
   formData.append("media", media);
-  return Api.post("/media/upload", formData, {
+  return Api.post("/admin/media/upload", formData, {
     headers: {
       "Content-Type": "multipart/form-data",
     },
@@ -98,17 +98,17 @@ export const getServices = function (
   limit: number = 10,
   type?: string
 ) {
-  return Api.get(`/service?page=${page}&limit=${limit}&type=${type}`);
+  return Api.get(`/admin/service?page=${page}&limit=${limit}&type=${type}`);
 };
 
 export const getService = function (id: string) {
-  return Api.get(`/service/show/${id}`);
+  return Api.get(`/admin/service/show/${id}`);
 };
 
 export const addService = function (body: any) {
-  return Api.post("/service/create", body);
+  return Api.post("/admin/service/create", body);
 };
 
 export const deleteService = function (id: string) {
-  return Api.delete(`/service/delete/${id}`);
+  return Api.delete(`/admin/service/delete/${id}`);
 };
